rewrite setup function

+++ inference-engine/tests/functional/plugin/shared/include/single_layer_tests/deformable_convolution.hpp
@@ -21,8 +21,8 @@ typedef std::tuple<
         std::vector<ptrdiff_t>,         // Pad end
         InferenceEngine::SizeVector,    // Dilation
         size_t,                         // Num out channels
-        int64_t,                        // Group
-        int64_t,                        // Deformable group
+	size_t,                         // Group
+	size_t,                         // Deformable group
 	ngraph::op::PadType             // Padding type
 > deformableConvSpecificParams;
 typedef std::tuple<
